stage4: fix Android test byte literals exceeding signed Byte range
C6.1 follow-up (append-only on remediation/stage-4-link-release). CI
:mesh:compileDebugUnitTestKotlin failed on two test-only byte literals:

  DeliveryTrackerTest.kt:281  ByteArray(16) { 0xA1 }
  DeliveryTrackerTest.kt:282  ByteArray(16) { 0xB1 }

`ByteArray(size, init)` expects `(Int) -> Byte`. Kotlin constant-narrows an
Int literal to Byte only when it fits in signed Byte (-128..127); 0xA1 (161)
and 0xB1 (177) are out of range, so they inferred Int and the test compile
rejected them ("Argument type mismatch: actual type is 'kotlin.Int', but
'kotlin.Byte' was expected"). In-range literals like 0x20 compiled fine.

Fixed with `0xA1.toByte()` / `0xB1.toByte()` (widened to Int then narrowed
explicitly), with an inline note. Production code (:mesh:compileDebugKotlin)
compiled clean -- this is test-only.

Verification:
  * ci/symbols.py: 68 Kotlin files, 0 unresolved.
  * ci/check_parity.py --scope repo: invariants A-H hold (passed=7, failed=0).
  * Android :mesh:testDebugUnitTest re-validated by CI on this commit.

+++ android/mesh/src/test/java/io/godstone/mesh/delivery/DeliveryTrackerTest.kt
@@ -278,8 +278,11 @@ class DeliveryTrackerTest {
         val journal = FakeJournal()
         val tracker = DeliveryTracker(journal, FakeAuthenticator(true))
         val mid = msgId(24)
-        val nodeA = ByteArray(16) { 0xA1 }
-        val nodeB = ByteArray(16) { 0xB1 }
+        // 0xA1 / 0xB1 exceed signed Byte (-128..127), so they must be widened
+        // to Int then narrowed with .toByte() -- Kotlin constant-narrows only
+        // in-range literals (e.g. 0x20), not out-of-range ones.
+        val nodeA = ByteArray(16) { 0xA1.toByte() }
+        val nodeB = ByteArray(16) { 0xB1.toByte() }
         assertEquals(EnqueueResult.Created, tracker.enqueue(mid, AckMode.SINGLE_RECIPIENT, nodeA))
         // Different intended recipient for the SAME msg_id -> ConflictRecipient.
         assertEquals(EnqueueResult.ConflictRecipient, tracker.enqueue(mid, AckMode.SINGLE_RECIPIENT, nodeB))
